move TestDataFileUtil to org.knora.webapi.util package
